+++ src/Compilers/BladeCompiler.php
@@ -9,6 +9,7 @@ use Weglot\Parser\ConfigProvider\ServerConfigProvider;
 use Weglot\Parser\Parser;
 use Weglot\Translate\Cache\LaravelCachePool;
 use Weglot\Translate\Compilers\Concerns\CompilesTranslations;
+use Weglot\Translate\TranslateServiceProvider;
 
 /**
  * Class BladeCompiler
@@ -33,17 +34,22 @@ class BladeCompiler extends LaravelBladeCompiler implements CompilerInterface
     public function compileString($value)
     {
         $contents = parent::compileString($value);
+        $contents = $this->cleanPhpLangTags($contents);
+
         $config = config('weglot-translate');
 
         $client = new Client($config['api_key']);
         $configProvider = new ServerConfigProvider();
+
+        // custom user-agent
+        $client->getHttpClient()->addUserAgentInfo('laravel', TranslateServiceProvider::VERSION);
 
         if ($config['cache']) {
             $cachePool = new LaravelCachePool();
             $client->setCacheItemPool($cachePool);
         }
 
-        $locale = weglotCurrentLocale();
+        $locale = $this->currentLocale();
         if ($locale !== $config['original_language']) {
             $parser = new Parser($client, $configProvider, $config['exclude_blocks']);
             $contents = $parser->translate($contents, $config['original_language'], $locale);
@@ -60,7 +66,12 @@ class BladeCompiler extends LaravelBladeCompiler implements CompilerInterface
      */
     public function getCompiledPath($path)
     {
-        $localizedPath = sprintf('%s|%s', weglotCurrentLocale(), $path);
+        $localizedPath = sprintf('%s|%s', $this->currentLocale(), $path);
         return $this->cachePath . '/' . sha1($localizedPath) . '.php';
+    }
+
+    private function currentLocale()
+    {
+        return weglotCurrentUrlInstance()->detectCurrentLanguage();
     }
 }