+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2018-05-07
+### Added
+- Custom user-agent injection
+- Method to find & parse `__()` & `trans()` PHP tags
+### Changed
+- Upgrading library version
+- Implementing Util.Url class
+
+## [0.1.2] - 2018-05-07
+### Added
+- README: Slack tag
+### Changed
+- PHP library version
+
 ## [0.1.1] - 2018-05-04
 ### Added
 - Compiler: Overwriting @lang behavior